refactor: Co-locate file upload and download tests
